Enforce maximum USDC transaction limit across services
