+++ src/constants.ts
@@ -2,6 +2,7 @@
 export const DEFAULT_OPTIONS = ["yes", "no"];
 export const DEFAULT_AMOUNT = "0.1";
 export const USDC_TOKEN_ADDRESS = "0x5dEaC602762362FE5f135FA5904351916053cF70";
+export const MAX_USDC_AMOUNT = 10; // Maximum allowed USDC transaction amount
 
 // Help message for users
 export const HELP_MESSAGE = `Available commands:
