+++ src/index.ts
@@ -8,8 +8,8 @@ import { TossManager } from "./toss-manager";
 import { handleCommand } from "./commands";
 import { WalletSendCallsCodec } from "@xmtp/content-type-wallet-send-calls";
 import { TransactionReferenceCodec } from "@xmtp/content-type-transaction-reference";
-import { verifyTransaction, extractERC20TransferData } from "@helpers/usdc";
-import { checkTransactionWithRetries } from "../helpers/transaction-checker";
+import {  checkTransactionWithRetries, extractERC20TransferData } from   "./transactions";
+
 
 /**
  * Custom JSON stringifier that can handle BigInt values
